Promote messageBox subtemplates as templates
