+++ packages/rocketchat-ui-message/client/messageBoxNotSubscribed.html
@@ -1,0 +1,25 @@
+<template name="messageBoxNotSubscribed" args="rid">
+	{{#if customTemplate}}
+		{{> Template.dynamic customTemplate }}
+	{{else}}
+		{{#if canJoinRoom}}
+			<div class="rc-message-box__join">
+				{{{_ "you_are_in_preview_mode_of" room_name=roomName}}}
+				{{#if isJoinCodeRequired}}
+					<input type="password" name="joinCode" placeholder="{{_ 'Password'}}" class="rc-input__element rc-message-box__join-code">
+				{{/if}}
+
+				<button class="rc-button rc-button--primary rc-button--small rc-message-box__join-button js-join">{{_ "join"}}</button>
+			</div>
+		{{/if}}
+
+		{{#if isAnonymousReadAllowed}}
+			<div class="rc-button-group">
+				<button class="rc-button rc-button--primary rc-button--small js-register">{{_ "Sign_in_to_start_talking"}}</button>
+				{{#if isAnonymousWriteAllowed}}
+					<button class="rc-button rc-button--small js-register-anonymous">{{_ "Or_talk_as_anonymous"}}</button>
+				{{/if}}
+			</div>
+		{{/if}}
+	{{/if}}
+</template>
