+++ packages/rocketchat-ui-message/client/messageBoxTyping.html
@@ -1,0 +1,20 @@
+<template name="messageBoxTyping" args="rid">
+	{{#with data}}
+		<div class="rc-message-box__typing">
+			<span class="rc-message-box__typing-user">{{users}}</span>
+			{{#if multi}}
+				{{#if selfTyping}}
+					{{_ "are_also_typing"}}
+				{{else}}
+					{{_ "are_typing"}}
+				{{/if}}
+			{{else}}
+				{{#if selfTyping}}
+					{{_ "is_also_typing" context="male"}}
+				{{else}}
+					{{_ "is_typing" context="male"}}
+				{{/if}}
+			{{/if}}
+		</div>
+	{{/with}}
+</template>
