+++ packages/rocketchat-ui-message/client/startup/index.js
@@ -1,0 +1,2 @@
+import './messageBoxActions';
+import './restoreText';
