+++ packages/rocketchat-ui-message/client/startup/restoreText.js
@@ -1,0 +1,15 @@
+import { Meteor } from 'meteor/meteor';
+import { RoomManager, chatMessages } from 'meteor/rocketchat:ui';
+
+
+Meteor.startup(() => {
+	RocketChat.callbacks.add('enter-room', () => {
+		setTimeout(() => {
+			if (!chatMessages[RoomManager.openedRoom].input) {
+				return;
+			}
+
+			chatMessages[RoomManager.openedRoom].restoreText(RoomManager.openedRoom);
+		}, 200);
+	});
+});
